requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
Pillow>=10.0.0
numpy>=1.26.0
scipy>=1.13.0
scikit-image>=0.24.0
opencv-python-headless>=4.10.0
fonttools>=4.53.0
PyMuPDF>=1.24.0
pytesseract>=0.3.10
